tools(crowtty): turn it into a lib
